pip-requirements-dev.txt
django-extensions
django-debug-toolbar
ipython
ipdb
fabric
pygraphviz
shortuuid
mock
django-nose==1.2
responses
simplejson
